test(cli): :wrench: Add possibility to create e2e sessions
Adds scripts to create a e2e test directory where the current coat cli source is installed into.

This is meant to be used for longer e2e testing sessions and enables to quickly watch e2e tests without constantly rebuilding the cli from the current source.

+++ packages/cli/test/utils/e2e-teardown.ts
@@ -6,6 +6,14 @@ export default async (): Promise<void> => {
     return;
   }
 
+  if (
+    process.env.COAT_CLI_E2E_SESSION_TARBALL_PATH &&
+    process.env.COAT_CLI_E2E_SESSION_TMP_PATH
+  ) {
+    // Skip teardown, since a test session has been created
+    return;
+  }
+
   const tmpDir = process.env.COAT_CLI_TMP_E2E_PATH;
   await fs.remove(tmpDir);
 };
